backend/requirements.txt
# Django Core
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.2.0
django-filter>=23.2
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# Async Processing
celery>=5.3.0
redis>=4.6.0
django-celery-results>=2.5.1
django-celery-beat>=2.5.0


# File Storage & Cloud
boto3>=1.28.0
google-cloud-storage>=2.10.0
django-storages>=1.14.0

# Image Processing & Medical Imaging
Pillow>=10.0.0
nibabel>=5.1.0
SimpleITK>=2.3.0
pydicom>=2.4.0

# ML Inference
torch>=2.0.0
torchvision>=0.15.0
monai>=1.3.0
onnx>=1.14.0
onnxruntime>=1.15.0

# 3D Visualization
trimesh>=3.22.0
pygltflib>=1.15.0
numpy>=1.24.0

# PDF Generation
reportlab>=4.0.0
weasyprint>=60.0

# Security & Authentication
cryptography>=41.0.0
PyJWT>=2.8.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0
python-multipart>=0.0.6
requests>=2.31.0

# API Documentation
drf-spectacular>=0.26.0

# Development
python-decouple>=3.8
ipython>=8.14.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Monitoring
sentry-sdk>=1.30.0
